pyproject.toml
[tool.poetry]
name = "orange-bot"
version = "0.1.0"
description = "internet surfer"
authors = ["kpomak <romka_ne@mail.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "orange_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "2.25.1"
python-dotenv = "1.0.0"
black = "23.3.0"
apscheduler = "3.10.1"
sqlalchemy = "2.0.12"
requests = "2.29.0"
pygithub = "1.58.1"
psycopg2-binary = "^2.9.6"
deepgram-sdk = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
